Move kontra game from separate app in public to a React component
Biggest change came in how we used custom fonts + TypeScript support (which is nice)
