--- public/games/index.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="manifest" href="manifest.json" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap"
-      rel="stylesheet"
-    />
-    <title>Plucky</title>
-    <style>
-      html,
-      body {
-        width: 100%;
-        height: 100%;
-        margin: 0;
-        padding: 0;
-      }
-
-      .game-container {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        flex-direction: column;
-        width: 100%;
-        height: 100%;
-      }
-
-      #game {
-        margin: 0 auto;
-      }
-
-      .press-start-2p-regular {
-        font-family: "Press Start 2P", system-ui;
-        font-weight: 400;
-        font-style: normal;
-      }
-    </style>
-  </head>
-
-  <body>
-    <div class="game-container">
-      <canvas id="game"></canvas>
-    </div>
-    <script src="lib/kontra.v10.0.2.js" lang="javascript"></script>
-    <script src="plucky/game.js" lang="javascript"></script>
-  </body>
-</html>
