--- public/games/manifest.json
@@ -1,6 +1,0 @@
-{
-  "short_name": "Plucky",
-  "name": "Plucky",
-  "display": "standalone",
-  "orientation": "portrait"
-}
